#180: Improved inheritance chain for remote metadata readers.
