+++ jdbc-adapter/virtualschema-jdbc-adapter/src/main/java/com/exasol/adapter/dialects/db2/DB2MetadataReader.java
@@ -3,17 +3,19 @@ package com.exasol.adapter.dialects.db2;
 import java.sql.Connection;
 
 import com.exasol.adapter.AdapterProperties;
-import com.exasol.adapter.jdbc.BaseRemoteMetadataReader;
-import com.exasol.adapter.jdbc.ColumnMetadataReader;
+import com.exasol.adapter.dialects.BaseIdentifierConverter;
+import com.exasol.adapter.dialects.IdentifierConverter;
+import com.exasol.adapter.jdbc.*;
 
 /**
  * This class reads and DB2-specific database metadata
  */
-public class DB2MetadataReader extends BaseRemoteMetadataReader {
+public class DB2MetadataReader extends AbstractRemoteMetadataReader {
     /**
      * Create a new instance of a {@link DB2MetadataReader}
      *
-     * @param connection database connection through which the reader retrieves the metadata from the remote source
+     * @param connection database connection through which the reader retrieves the metadata from the remote
+     *                   source
      * @param properties user-defined properties
      */
     public DB2MetadataReader(final Connection connection, final AdapterProperties properties) {
@@ -23,5 +25,16 @@ public class DB2MetadataReader extends BaseRemoteMetadataReader {
     @Override
     protected ColumnMetadataReader createColumnMetadataReader() {
         return new DB2ColumnMetadataReader(this.connection, this.properties, getIdentifierConverter());
+    }
+
+    @Override
+    protected TableMetadataReader createTableMetadataReader() {
+        return new BaseTableMetadataReader(this.connection, this.columnMetadataReader, this.properties,
+                this.identifierConverter);
+    }
+
+    @Override
+    protected IdentifierConverter createIdentifierConverter() {
+        return BaseIdentifierConverter.createDefault();
     }
 }
