+++ jdbc-adapter/virtualschema-jdbc-adapter/src/main/java/com/exasol/adapter/dialects/exasol/ExasolMetadataReader.java
@@ -1,19 +1,21 @@
 package com.exasol.adapter.dialects.exasol;
 
-import com.exasol.adapter.AdapterProperties;
-import com.exasol.adapter.jdbc.BaseRemoteMetadataReader;
-import com.exasol.adapter.jdbc.ColumnMetadataReader;
-
 import java.sql.Connection;
+
+import com.exasol.adapter.AdapterProperties;
+import com.exasol.adapter.dialects.BaseIdentifierConverter;
+import com.exasol.adapter.dialects.IdentifierConverter;
+import com.exasol.adapter.jdbc.*;
 
 /**
  * This class reads and Exasol-specific database metadata
  */
-public class ExasolMetadataReader extends BaseRemoteMetadataReader {
+public class ExasolMetadataReader extends AbstractRemoteMetadataReader {
     /**
      * Create a new instance of a {@link ExasolMetadataReader}
      *
-     * @param connection database connection through which the reader retrieves the metadata from the remote source
+     * @param connection database connection through which the reader retrieves the metadata from the remote
+     *                   source
      * @param properties user-defined properties
      */
     public ExasolMetadataReader(final Connection connection, final AdapterProperties properties) {
@@ -23,5 +25,16 @@ public class ExasolMetadataReader extends BaseRemoteMetadataReader {
     @Override
     protected ColumnMetadataReader createColumnMetadataReader() {
         return new ExasolColumnMetadataReader(this.connection, this.properties, getIdentifierConverter());
+    }
+
+    @Override
+    protected TableMetadataReader createTableMetadataReader() {
+        return new BaseTableMetadataReader(this.connection, this.columnMetadataReader, this.properties,
+                this.identifierConverter);
+    }
+
+    @Override
+    protected IdentifierConverter createIdentifierConverter() {
+        return BaseIdentifierConverter.createDefault();
     }
 }
