+++ jdbc-adapter/virtualschema-jdbc-adapter/src/main/java/com/exasol/adapter/dialects/generic/GenericMetadataReader.java
@@ -4,16 +4,17 @@ import java.sql.Connection;
 
 import com.exasol.adapter.AdapterProperties;
 import com.exasol.adapter.dialects.IdentifierConverter;
-import com.exasol.adapter.jdbc.BaseRemoteMetadataReader;
+import com.exasol.adapter.jdbc.*;
 
 /**
  * This class reads Generic database metadata
  */
-public class GenericMetadataReader extends BaseRemoteMetadataReader {
+public class GenericMetadataReader extends AbstractRemoteMetadataReader {
     /**
      * Create a new instance of a {@link GenericMetadataReader}
      *
-     * @param connection database connection through which the reader retrieves the metadata from the remote source
+     * @param connection database connection through which the reader retrieves the metadata from the remote
+     *                   source
      * @param properties user-defined properties
      */
     public GenericMetadataReader(final Connection connection, final AdapterProperties properties) {
@@ -23,5 +24,16 @@ public class GenericMetadataReader extends BaseRemoteMetadataReader {
     @Override
     protected IdentifierConverter createIdentifierConverter() {
         return new GenericIdentifierConverter(getSchemaAdapterNotes());
+    }
+
+    @Override
+    protected ColumnMetadataReader createColumnMetadataReader() {
+        return new BaseColumnMetadataReader(this.connection, this.properties, this.identifierConverter);
+    }
+
+    @Override
+    protected TableMetadataReader createTableMetadataReader() {
+        return new BaseTableMetadataReader(this.connection, this.columnMetadataReader, this.properties,
+                this.identifierConverter);
     }
 }
