+++ jdbc-adapter/virtualschema-jdbc-adapter/src/main/java/com/exasol/adapter/dialects/hive/HiveMetadataReader.java
@@ -5,17 +5,17 @@ import java.sql.Connection;
 import com.exasol.adapter.AdapterProperties;
 import com.exasol.adapter.dialects.BaseIdentifierConverter;
 import com.exasol.adapter.dialects.IdentifierConverter;
-import com.exasol.adapter.jdbc.BaseRemoteMetadataReader;
-import com.exasol.adapter.jdbc.IdentifierCaseHandling;
+import com.exasol.adapter.jdbc.*;
 
 /**
  * This class reads Hive-specific database metadata
  */
-public class HiveMetadataReader extends BaseRemoteMetadataReader {
+public class HiveMetadataReader extends AbstractRemoteMetadataReader {
     /**
      * Create a new instance of a {@link HiveMetadataReader}
      *
-     * @param connection database connection through which the reader retrieves the metadata from the remote source
+     * @param connection database connection through which the reader retrieves the metadata from the remote
+     *                   source
      * @param properties user-defined properties
      */
     public HiveMetadataReader(final Connection connection, final AdapterProperties properties) {
@@ -26,5 +26,16 @@ public class HiveMetadataReader extends BaseRemoteMetadataReader {
     protected IdentifierConverter createIdentifierConverter() {
         return new BaseIdentifierConverter(IdentifierCaseHandling.INTERPRET_AS_LOWER,
                 IdentifierCaseHandling.INTERPRET_AS_LOWER);
+    }
+
+    @Override
+    protected ColumnMetadataReader createColumnMetadataReader() {
+        return new BaseColumnMetadataReader(this.connection, this.properties, this.identifierConverter);
+    }
+
+    @Override
+    protected TableMetadataReader createTableMetadataReader() {
+        return new BaseTableMetadataReader(this.connection, this.columnMetadataReader, this.properties,
+                this.identifierConverter);
     }
 }
