+++ jdbc-adapter/virtualschema-jdbc-adapter/src/main/java/com/exasol/adapter/dialects/impala/ImpalaMetadataReader.java
@@ -10,11 +10,12 @@ import com.exasol.adapter.jdbc.*;
 /**
  * This class reads Impala-specific database metadata
  */
-public class ImpalaMetadataReader extends BaseRemoteMetadataReader {
+public class ImpalaMetadataReader extends AbstractRemoteMetadataReader {
     /**
      * Create a new instance of a {@link ImpalaMetadataReader}
      *
-     * @param connection database connection through which the reader retrieves the metadata from the remote source
+     * @param connection database connection through which the reader retrieves the metadata from the remote
+     *                   source
      * @param properties user-defined properties
      */
     public ImpalaMetadataReader(final Connection connection, final AdapterProperties properties) {
@@ -24,6 +25,17 @@ public class ImpalaMetadataReader extends BaseRemoteMetadataReader {
     @Override
     protected IdentifierConverter createIdentifierConverter() {
         return new BaseIdentifierConverter(IdentifierCaseHandling.INTERPRET_AS_LOWER,
-              IdentifierCaseHandling.INTERPRET_AS_LOWER);
+                IdentifierCaseHandling.INTERPRET_AS_LOWER);
+    }
+
+    @Override
+    protected ColumnMetadataReader createColumnMetadataReader() {
+        return new BaseColumnMetadataReader(this.connection, this.properties, this.identifierConverter);
+    }
+
+    @Override
+    protected TableMetadataReader createTableMetadataReader() {
+        return new BaseTableMetadataReader(this.connection, this.columnMetadataReader, this.properties,
+                this.identifierConverter);
     }
 }
