+++ jdbc-adapter/virtualschema-jdbc-adapter/src/main/java/com/exasol/adapter/dialects/oracle/OracleMetadataReader.java
@@ -3,16 +3,19 @@ package com.exasol.adapter.dialects.oracle;
 import java.sql.Connection;
 
 import com.exasol.adapter.AdapterProperties;
+import com.exasol.adapter.dialects.BaseIdentifierConverter;
+import com.exasol.adapter.dialects.IdentifierConverter;
 import com.exasol.adapter.jdbc.*;
 
 /**
  * This class reads and Oracle-specific database metadata
  */
-public class OracleMetadataReader extends BaseRemoteMetadataReader {
+public class OracleMetadataReader extends AbstractRemoteMetadataReader {
     /**
      * Create a new instance of a {@link OracleMetadataReader}
      *
-     * @param connection database connection through which the reader retrieves the metadata from the remote source
+     * @param connection database connection through which the reader retrieves the metadata from the remote
+     *                   source
      * @param properties user-defined properties
      */
     public OracleMetadataReader(final Connection connection, final AdapterProperties properties) {
@@ -28,5 +31,10 @@ public class OracleMetadataReader extends BaseRemoteMetadataReader {
     @Override
     protected ColumnMetadataReader createColumnMetadataReader() {
         return new OracleColumnMetadataReader(this.connection, this.properties, getIdentifierConverter());
+    }
+
+    @Override
+    protected IdentifierConverter createIdentifierConverter() {
+        return BaseIdentifierConverter.createDefault();
     }
 }
