+++ jdbc-adapter/virtualschema-jdbc-adapter/src/main/java/com/exasol/adapter/dialects/sqlserver/SqlServerMetadataReader.java
@@ -3,13 +3,14 @@ package com.exasol.adapter.dialects.sqlserver;
 import java.sql.Connection;
 
 import com.exasol.adapter.AdapterProperties;
-import com.exasol.adapter.jdbc.BaseRemoteMetadataReader;
-import com.exasol.adapter.jdbc.ColumnMetadataReader;
+import com.exasol.adapter.dialects.BaseIdentifierConverter;
+import com.exasol.adapter.dialects.IdentifierConverter;
+import com.exasol.adapter.jdbc.*;
 
 /**
  * This class implement a SQLServer-specific metadata reader
  */
-public class SqlServerMetadataReader extends BaseRemoteMetadataReader {
+public class SqlServerMetadataReader extends AbstractRemoteMetadataReader {
     /**
      * Create a new instance of a {@link SqlServerMetadataReader}
      *
@@ -23,5 +24,16 @@ public class SqlServerMetadataReader extends BaseRemoteMetadataReader {
     @Override
     protected ColumnMetadataReader createColumnMetadataReader() {
         return new SqlServerColumnMetadataReader(this.connection, this.properties, getIdentifierConverter());
+    }
+
+    @Override
+    protected TableMetadataReader createTableMetadataReader() {
+        return new BaseTableMetadataReader(this.connection, this.columnMetadataReader, this.properties,
+                this.identifierConverter);
+    }
+
+    @Override
+    protected IdentifierConverter createIdentifierConverter() {
+        return BaseIdentifierConverter.createDefault();
     }
 }
