+++ jdbc-adapter/virtualschema-jdbc-adapter/src/main/java/com/exasol/adapter/dialects/teradata/TeradataMetadataReader.java
@@ -3,17 +3,19 @@ package com.exasol.adapter.dialects.teradata;
 import java.sql.Connection;
 
 import com.exasol.adapter.AdapterProperties;
-import com.exasol.adapter.jdbc.BaseRemoteMetadataReader;
-import com.exasol.adapter.jdbc.ColumnMetadataReader;
+import com.exasol.adapter.dialects.BaseIdentifierConverter;
+import com.exasol.adapter.dialects.IdentifierConverter;
+import com.exasol.adapter.jdbc.*;
 
 /**
  * This class reads and Teradata-specific database metadata
  */
-public class TeradataMetadataReader extends BaseRemoteMetadataReader {
+public class TeradataMetadataReader extends AbstractRemoteMetadataReader {
     /**
      * Create a new instance of a {@link TeradataMetadataReader}
      *
-     * @param connection database connection through which the reader retrieves the metadata from the remote source
+     * @param connection database connection through which the reader retrieves the metadata from the remote
+     *                   source
      * @param properties user-defined properties
      */
     public TeradataMetadataReader(final Connection connection, final AdapterProperties properties) {
@@ -23,5 +25,16 @@ public class TeradataMetadataReader extends BaseRemoteMetadataReader {
     @Override
     protected ColumnMetadataReader createColumnMetadataReader() {
         return new TeradataColumnMetadataReader(this.connection, this.properties, getIdentifierConverter());
+    }
+
+    @Override
+    protected TableMetadataReader createTableMetadataReader() {
+        return new BaseTableMetadataReader(this.connection, this.columnMetadataReader, this.properties,
+                this.identifierConverter);
+    }
+
+    @Override
+    protected IdentifierConverter createIdentifierConverter() {
+        return BaseIdentifierConverter.createDefault();
     }
 }
